Убраны компоненты Offers и Destinations - теперь они часть формы
